K8s reporter to own its artifacts sub directory
The K8s reporter removes the artifacts directory whenver it starts the
test and it doesn't create it later if it failed to generate a client
for the cluster. Since other reporters also depend on the shared
artifacts dir, they fail as well.

This commit gives the K8s reporter it's own artifacts sub dir to own and
manage however it needs to.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>